+++ spring-data-eclipse-store/src/main/java/software/xdev/spring/data/eclipse/store/repository/interfaces/EclipseStoreRepository.java
@@ -23,7 +23,7 @@ import org.springframework.data.repository.NoRepositoryBean;
 public interface EclipseStoreRepository<T, ID>
 	extends
 	EclipseStoreListCrudRepository<T, ID>,
-	EclipseStoreListPagingAndSortingRepositoryRepository<T, ID>,
+	EclipseStoreListPagingAndSortingRepository<T, ID>,
 	EclipseStoreQueryByExampleExecutor<T>
 {
 }