+++ spring-data-eclipse-store/src/main/java/software/xdev/spring/data/eclipse/store/repository/support/SimpleEclipseStoreRepository.java
@@ -34,12 +34,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.repository.query.FluentQuery;
 
-import software.xdev.spring.data.eclipse.store.repository.EclipseStoreStorage;
+import software.xdev.spring.data.eclipse.store.repository.StorageCommunicator;
 import software.xdev.spring.data.eclipse.store.repository.access.AccessHelper;
 import software.xdev.spring.data.eclipse.store.repository.interfaces.EclipseStoreCrudRepository;
 import software.xdev.spring.data.eclipse.store.repository.interfaces.EclipseStoreListCrudRepository;
-import software.xdev.spring.data.eclipse.store.repository.interfaces.EclipseStoreListPagingAndSortingRepositoryRepository;
-import software.xdev.spring.data.eclipse.store.repository.interfaces.EclipseStorePagingAndSortingRepositoryRepository;
+import software.xdev.spring.data.eclipse.store.repository.interfaces.EclipseStoreListPagingAndSortingRepository;
+import software.xdev.spring.data.eclipse.store.repository.interfaces.EclipseStorePagingAndSortingRepository;
 import software.xdev.spring.data.eclipse.store.repository.interfaces.EclipseStoreQueryByExampleExecutor;
 import software.xdev.spring.data.eclipse.store.repository.interfaces.EclipseStoreRepository;
 import software.xdev.spring.data.eclipse.store.repository.query.by.example.EclipseStoreFetchableFluentQuery;
@@ -61,21 +61,21 @@ import software.xdev.spring.data.eclipse.store.transactions.EclipseStoreTransact
 public class SimpleEclipseStoreRepository<T, ID>
 	implements
 	EclipseStoreRepository<T, ID>,
-	EclipseStorePagingAndSortingRepositoryRepository<T, ID>,
-	EclipseStoreListPagingAndSortingRepositoryRepository<T, ID>,
+	EclipseStorePagingAndSortingRepository<T, ID>,
+	EclipseStoreListPagingAndSortingRepository<T, ID>,
 	EclipseStoreCrudRepository<T, ID>,
 	EclipseStoreListCrudRepository<T, ID>,
 	EclipseStoreQueryByExampleExecutor<T>
 {
 	private static final Logger LOG = LoggerFactory.getLogger(SimpleEclipseStoreRepository.class);
-	private final EclipseStoreStorage storage;
+	private final StorageCommunicator storage;
 	private final Class<T> domainClass;
 	private final WorkingCopier<T> copier;
 	private final EclipseStoreTransactionManager transactionManager;
 	private final IdManager<T, ID> idManager;
 	
 	public SimpleEclipseStoreRepository(
-		final EclipseStoreStorage storage,
+		final StorageCommunicator storage,
 		final WorkingCopier<T> copier,
 		final Class<T> domainClass,
 		final EclipseStoreTransactionManager transactionManager,
@@ -398,5 +398,10 @@ public class SimpleEclipseStoreRepository<T, ID>
 		return this.storage.getReadWriteLock().read(
 			() -> queryFunction.apply(query)
 		);
+	}
+	
+	public boolean isLazy()
+	{
+		return false;
 	}
 }