+++ spring-data-eclipse-store/src/main/java/software/xdev/spring/data/eclipse/store/repository/support/copier/working/RecursiveWorkingCopier.java
@@ -28,8 +28,11 @@ import java.util.Optional;
 import java.util.TreeMap;
 import java.util.TreeSet;
 
+import jakarta.validation.Validator;
+
 import org.eclipse.serializer.reference.Lazy;
 import org.eclipse.serializer.reference.ObjectSwizzling;
+import org.eclipse.serializer.reflect.ClassLoaderProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -72,15 +75,29 @@ public class RecursiveWorkingCopier<T> implements WorkingCopier<T>
 		final VersionManagerProvider versionManagerProvider,
 		final PersistableChecker persistableChecker,
 		final SupportedChecker supportedChecker,
-		final ObjectSwizzling objectSwizzling
+		final ObjectSwizzling objectSwizzling,
+		final Validator validator,
+		final ClassLoaderProvider currentClassLoaderProvider
 	)
 	{
 		this.domainClass = domainClass;
 		this.registry = registry;
 		this.workingCopyToStorageCopier =
-			new RegisteringWorkingCopyToStorageCopier(registry, supportedChecker, objectSwizzling, this);
+			new RegisteringWorkingCopyToStorageCopier(
+				registry,
+				supportedChecker,
+				objectSwizzling,
+				this,
+				validator,
+				currentClassLoaderProvider);
 		this.storageToWorkingCopyCopier =
-			new RegisteringStorageToWorkingCopyCopier(registry, supportedChecker, objectSwizzling, this);
+			new RegisteringStorageToWorkingCopyCopier(
+				registry,
+				supportedChecker,
+				objectSwizzling,
+				this,
+				validator,
+				currentClassLoaderProvider);
 		this.idManagerProvider = idManagerProvider;
 		this.versionManagerProvider = versionManagerProvider;
 		this.persistableChecker = persistableChecker;